rename: 서비스 계층 패키지 구조 변경 (application → application.service)
