+++ admin/src/main/java/kernel360/ckt/admin/application/service/command/UpdateVehicleCommand.java
@@ -1,6 +1,5 @@
-package kernel360.ckt.admin.application.command;
+package kernel360.ckt.admin.application.service.command;
 
-import kernel360.ckt.core.domain.entity.VehicleEntity;
 import kernel360.ckt.core.domain.enums.FuelType;
 import kernel360.ckt.core.domain.enums.TransmissionType;
 import kernel360.ckt.core.domain.enums.VehicleStatus;
@@ -9,8 +8,7 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class CreateVehicleCommand {
-    private final String registrationNumber;
+public class UpdateVehicleCommand {
     private final String modelYear;
     private final String manufacturer;
     private final String modelName;
@@ -19,18 +17,4 @@ public class CreateVehicleCommand {
     private final TransmissionType transmissionType;
     private final VehicleStatus status;
     private final String memo;
-
-    public VehicleEntity toEntity() {
-        return VehicleEntity.create(
-            this.registrationNumber,
-            this.modelYear,
-            this.manufacturer,
-            this.modelName,
-            this.batteryVoltage,
-            this.fuelType,
-            this.transmissionType,
-            this.status,
-            this.memo
-        );
-    }
 }
