+++ admin/src/main/java/kernel360/ckt/admin/ui/VehicleControlTowerController.java
@@ -1,15 +1,12 @@
 package kernel360.ckt.admin.ui;
 
 import kernel360.ckt.admin.ui.dto.response.ControlTowerSummaryResponse;
-import kernel360.ckt.admin.application.VehicleService;
-import kernel360.ckt.admin.ui.dto.response.GpsPointResponse;
+import kernel360.ckt.admin.application.service.VehicleService;
 import kernel360.ckt.admin.ui.dto.response.RunningVehicleResponse;
 import kernel360.ckt.core.common.response.CommonResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
