+++ admin/src/main/java/kernel360/ckt/admin/ui/VehicleController.java
@@ -1,8 +1,8 @@
 package kernel360.ckt.admin.ui;
 
-import kernel360.ckt.admin.application.VehicleService;
-import kernel360.ckt.admin.application.command.CreateVehicleCommand;
-import kernel360.ckt.admin.application.command.UpdateVehicleCommand;
+import kernel360.ckt.admin.application.service.VehicleService;
+import kernel360.ckt.admin.application.service.command.CreateVehicleCommand;
+import kernel360.ckt.admin.application.service.command.UpdateVehicleCommand;
 import kernel360.ckt.admin.ui.dto.request.VehicleCreateRequest;
 import kernel360.ckt.admin.ui.dto.request.VehicleUpdateRequest;
 import kernel360.ckt.admin.ui.dto.response.*;
